move model init's into their own model module
